requirements.txt
Django==4.0.4
jsonpickle==2.2.0
PyMySQL==1.0.2
pytest==7.1.2
requests==2.28.0
selenium==4.4.3
websocket_client==1.4.1
xlrd==2.0.1
xlwings==0.27.8
